feat(casper): coin module cleanup and tests part 2
