+++ libs/coin-modules/coin-casper/.unimportedrc.json
@@ -2,14 +2,15 @@
   "entry": [
     "src/api/index.ts",
     "src/bridge/index.ts",
-    "src/bridge/deviceTransactionConfig.ts",
-    "src/common-logic/index.ts",
+    "src/deviceTransactionConfig.ts",
+    "src/logic/index.ts",
     "src/signer/index.ts",
     "src/test/index.ts",
     "src/types/index.ts"
   ],
   "ignorePatterns": [
     "**/node_modules/**",
+    "**/__tests__/**",
     "**/*.fixture.ts",
     "**/*.mock.ts",
     "**/*.test.{js,jsx,ts,tsx}"
