+++ libs/coin-modules/coin-casper/src/__tests__/fixtures/addresses.fixture.ts
@@ -1,0 +1,18 @@
+export const TEST_ADDRESSES = {
+  SECP256K1: "0202ba6dc98cbe677711a45bf028a03646f9e588996eb223fad2485e8bc391b01581",
+  RECIPIENT_SECP256K1: "0203A17118eC0e64c4e4FdbDbEe0eA14D118C9aAf08C6c81bbB776Cae607cEB84EcB",
+  RECIPIENT_ED25519: "01e28f293af356e7a15068e535c248ec07c887b2ab7a5d9557037a0e998e5d97bf",
+  INVALID: "notavalidaddress",
+};
+
+export const TEST_TRANSACTION_HASHES = {
+  VALID: "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef",
+  PENDING: "pending0123456789abcdef0123456789abcdef0123456789abcdef01234567",
+  FAILED: "failed00123456789abcdef0123456789abcdef0123456789abcdef01234567",
+};
+
+export const TEST_TRANSFER_IDS = {
+  VALID: "123456",
+  NUMERIC: "654321",
+  INVALID: "not-a-valid-id",
+};
