+++ libs/coin-modules/coin-casper/src/__tests__/fixtures/index.ts
@@ -1,0 +1,5 @@
+export * from "./account.fixture";
+export * from "./addresses.fixture";
+export * from "./operation.fixture";
+export * from "./sync.fixture";
+export * from "./transaction.fixture";
