+++ libs/coin-modules/coin-casper/src/bridge/signOperation.ts
@@ -3,8 +3,8 @@ import { log } from "@ledgerhq/logs";
 import { Account, AccountBridge } from "@ledgerhq/types-live";
 import { KeyAlgorithm } from "casper-js-sdk";
 import { Observable } from "rxjs";
-import { getAddress } from "../bridge/bridgeHelpers/addresses";
-import { createNewTransaction } from "../bridge/bridgeHelpers/txn";
+import { createNewTransaction } from "../logic/craftTransaction";
+import { getAddress } from "../logic/validateAddress";
 import { Transaction } from "../types";
 import { CasperSigner } from "../types";
 import { buildOptimisticOperation } from "./buildOptimisticOperation";
