+++ libs/coin-modules/coin-casper/src/bridge/sync.ts
@@ -3,9 +3,9 @@ import { GetAccountShape } from "@ledgerhq/ledger-wallet-framework/bridge/jsHelp
 import { log } from "@ledgerhq/logs";
 import BigNumber from "bignumber.js";
 import flatMap from "lodash/flatMap";
-import { fetchBalance, fetchBlockHeight, fetchAccountStateInfo, fetchTxs } from "../../network/api";
-import { ITxnHistoryData } from "../../types/network";
-import { mapTxToOps } from "./txn";
+import { fetchBalance, fetchBlockHeight, fetchAccountStateInfo, fetchTxs } from "../network/api";
+import { ITxnHistoryData } from "../types/network";
+import { mapTxToOps } from "../logic/listOperations";
 
 export const getAccountShape: GetAccountShape = async info => {
   const { address, currency, derivationMode } = info;
