--- libs/coin-modules/coin-casper/src/common-logic/index.ts
@@ -1,10 +1,0 @@
-export {
-  isNoErrorReturnCode,
-  getPath,
-  isValidHex,
-  isValidBase64,
-  isError,
-  methodToString,
-  getBufferFromString,
-  getRandomTransferID,
-} from "./utils";
