+++ libs/coin-modules/coin-casper/src/config.ts
@@ -1,11 +1,4 @@
-import { CurrencyConfig } from "@ledgerhq/coin-module-framework/config";
-
-export type CasperCoinConfig = () => CurrencyConfig & {
-  infra: {
-    API_CASPER_NODE_ENDPOINT: string;
-    API_CASPER_INDEXER: string;
-  };
-};
+import { CasperCoinConfig } from "./types/config";
 
 let coinConfig: CasperCoinConfig | undefined;
