+++ libs/coin-modules/coin-casper/src/logic/estimateFees.ts
@@ -1,5 +1,5 @@
 import BigNumber from "bignumber.js";
-import { CASPER_FEES_MOTES } from "../../constants";
+import { CASPER_FEES_MOTES } from "../constants";
 
 export function getEstimatedFees(): BigNumber {
   return new BigNumber(CASPER_FEES_MOTES);
