+++ libs/coin-modules/coin-casper/src/logic/listOperations.ts
@@ -1,17 +1,13 @@
 import { getCryptoCurrencyById } from "@ledgerhq/ledger-wallet-framework/currencies";
-import { InvalidAddress } from "@ledgerhq/ledger-wallet-framework/errors";
 import { encodeOperationId } from "@ledgerhq/ledger-wallet-framework/operation";
-import { log } from "@ledgerhq/logs";
 import { Unit } from "@ledgerhq/ledger-wallet-framework/types";
+import { log } from "@ledgerhq/logs";
 import BigNumber from "bignumber.js";
-import { CasperNetwork, PublicKey, Transaction } from "casper-js-sdk";
 import invariant from "invariant";
-import { getCasperNodeRpcClient } from "../../network/api";
-import { ITxnHistoryData } from "../../types/network";
-import { CASPER_DEFAULT_TTL, CASPER_NETWORK } from "../../constants";
-import { CasperOperation } from "../../types";
-import { casperAccountHashFromPublicKey, isAddressValid } from "./addresses";
-import { getEstimatedFees } from "./fee";
+import { ITxnHistoryData } from "../types/network";
+import { CasperOperation } from "../types";
+import { getEstimatedFees } from "./estimateFees";
+import { casperAccountHashFromPublicKey } from "./validateAddress";
 
 export const getUnit = (): Unit => getCryptoCurrencyById("casper").units[0];
 
@@ -89,33 +85,3 @@ export function mapTxToOps(
     }
   };
 }
-
-export const createNewTransaction = async (
-  sender: string,
-  recipient: string,
-  amount: BigNumber,
-  fees: BigNumber,
-  transferId?: string,
-  network = CASPER_NETWORK,
-): Promise<Transaction> => {
-  log("debug", `Creating new Transaction: ${sender}, ${recipient}, ${network}`);
-
-  if (recipient && !isAddressValid(recipient)) {
-    throw new InvalidAddress(`Invalid recipient Address ${recipient}`);
-  }
-
-  const client = getCasperNodeRpcClient();
-  const helper = await CasperNetwork.create(client);
-
-  const tx = helper.createTransferTransaction(
-    PublicKey.fromHex(sender),
-    PublicKey.fromHex(recipient),
-    network,
-    amount.toString(),
-    fees.toNumber(),
-    CASPER_DEFAULT_TTL,
-    parseInt(transferId ?? "0"),
-  );
-
-  return tx;
-};
