+++ libs/coin-modules/coin-casper/src/test/bridgeDatasetTest.ts
@@ -5,7 +5,7 @@ import {
 } from "@ledgerhq/ledger-wallet-framework/errors";
 import type { CurrenciesData, DatasetTest } from "@ledgerhq/types-live";
 import BigNumber from "bignumber.js";
-import { getEstimatedFees } from "../bridge/bridgeHelpers/fee";
+import { getEstimatedFees } from "../logic/estimateFees";
 import { fromTransactionRaw } from "../bridge/transaction";
 import { CasperInvalidTransferId, InvalidMinimumAmount, MayBlockAccount } from "../errors";
 import type { Transaction } from "../types";
