+++ libs/coin-modules/coin-casper/src/test/index.ts
@@ -1,2 +1,1 @@
 export * from "./bridgeDatasetTest";
-export * from "./fixtures";
