+++ libs/coin-modules/coin-casper/src/types/bridge.ts
