+++ libs/coin-modules/coin-casper/src/types/config.ts
@@ -1,0 +1,8 @@
+import { CurrencyConfig } from "@ledgerhq/coin-module-framework/config";
+
+export type CasperCoinConfig = () => CurrencyConfig & {
+  infra: {
+    API_CASPER_NODE_ENDPOINT: string;
+    API_CASPER_INDEXER: string;
+  };
+};
