+++ libs/coin-modules/coin-casper/src/types/index.ts
@@ -1,3 +1,4 @@
-export * from "./common";
+export * from "./bridge";
+export * from "./config";
 export * from "./network";
 export * from "./signer";
